brain-gaze-server/requirements.txt
Flask
flask_cors
pillow
